% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelCurrentStatus.R
\name{ccl.fit}
\alias{ccl.fit}
\title{Calcuate the scaled coefficients from the conditional censoring logistic (CCL) estimator}
\usage{
ccl.fit(delta, ctime, predictors, n.ptb, seed = 1)
}
\arguments{
\item{delta}{matrix of event indicator}

\item{ctime}{matrix of censoring/monitoring time}

\item{predictors}{matrix of predictors}

\item{n.ptb}{number of perturbations}

\item{seed}{random seed for resampling}
}
\description{
Calcuate the scaled coefficients from the conditional censoring logistic (CCL) estimator
}
