% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelCurrentStatus.R
\name{ccl.fit}
\alias{ccl.fit}
\title{...}
\usage{
ccl.fit(delta, ctime, predictors, n.ptb, seed = 1)
}
\arguments{
\item{delta}{...}

\item{ctime}{...}

\item{predictors}{...}

\item{n.ptb}{...}

\item{seed}{...}
}
\description{
...
}
