% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelCurrentStatus.R
\name{ccl.fit}
\alias{ccl.fit}
\title{Calcuate the scaled coefficents from the conditional censoring logistic (CCL) estimator}
\usage{
ccl.fit(delta, ctime, predictors, n.ptb, seed = 1)
}
\arguments{
\item{delta}{event indicator matrix}

\item{ctime}{censoring time matrix}

\item{predictors}{predictor matrix}

\item{n.ptb}{number of perturbations}

\item{seed}{random seed for resampling}
}
\description{
Calcuate the scaled coefficents from the conditional censoring logistic (CCL) estimator
}
