% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelCurrentStatus.R
\name{ccl.roc}
\alias{ccl.roc}
\title{Evaluating the model metrics from ROC curve via kernel smoothing}
\usage{
ccl.roc(data, fit, t0, h)
}
\arguments{
\item{data}{data matrix in 'long' format}

\item{fit}{results from function ccl.fit}

\item{t0}{pre-specified prediction time}

\item{h}{bandwidth for kernel smoothing}
}
\description{
Evaluating the model metrics from ROC curve via kernel smoothing
}
